API/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
django-ratelimit>=4.1.0
pillow>=10.0
requests>=2.32.0
drf-spectacular>=0.26
